OCL initialization
All loading commented out - these will be dynamically loaded when a file
is selected using OCL -> Load Document (because all references to these
files are done via `platform:/plugin`. Loading them at this point
doesn't work because of a timing issue? The plugin may not be fully
registered?  In any case, OCL -> Load Document works - and is desirable
for now.