Add generalized Acceleo calls to IndexUtilities.
ECrossReferenceAdapterCrossReferenceProvider and
ResourceSetRootEObjectProvider are cloned from MDoc

NOTE: For some reason the ResourceSet in my tests did not list every
resource in the XText project I was editing while testing this.

NOTE: Cache QueryEnvironments so we don't keep recreating them. When doing that, makepackage registration part of what is cached.